Moved Confirm to components folder
