+++ src/components/confirmationModal/Confirm.scss
